convert heat templates into disk images via heat
Use Heat to convert a heat template into disk images unique to each
server.  The script uses diskimage-builder to create images for each
server and then uploads them to the local glance.  At conclusion, it
saves an updated template with references to the metadata removed
and the proper image references.